requirements.txt
appnope==0.1.0
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cycler==0.10.0
decorator==4.1.2
docopt==0.6.2
entrypoints==0.2.3
Flask==0.12.2
Flask-API==0.7.1
Flask-Cors==3.0.3
get==0.0.21
html5lib==0.9999999
idna==2.5
itsdangerous==0.24
jedi==0.10.2
lxml==3.8.0
Markdown==2.6.8
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
nbconvert==5.2.1
nbformat==4.3.0
notebook==5.0.0
numpy==1.13.1
olefile==0.44
pandocfilters==1.4.2
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
pipreqs==0.4.9
post==0.0.13
prompt-toolkit==1.0.15
protobuf==3.3.0
ptyprocess==0.5.2
public==0.0.38
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.1
query-string==0.0.12
request==0.0.13
requests==2.18.2
setupfiles==0.0.50
simplegeneric==0.8.1
six==1.10.0
tensorflow==1.3.0
terminado==0.6
testpath==0.3.1
tornado==4.5.1
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.0.0
yarg==0.1.9
